+++ CHANGES.rst
@@ -3,7 +3,7 @@
 Changelog
 =========
 
-Version 0.9.0 - 0.9.3
+Version 0.9.0 - 0.9.4
 ---------------------
 - Annoying bug fixed.
 
@@ -20,6 +20,8 @@ Version 0.9.0 - 0.9.3
 - Add `crumb ls` and `crumb intersect` CLI.
 
 - Add `is_crumb_arg` function to utils.
+
+- Fix error in `crumb copy` parameter names.
 
 
 Version 0.8.0 - 0.8.3